subscribe to setting action
