+++ client/app.js
@@ -10,6 +10,7 @@ import { push, routerMiddleware, syncHistoryWithStore } from 'react-router-redux
 import { loadCredentials } from './actions/auth';
 import routes from './routes';
 import configureStore from './store/configureStore';
+import * as constants from './constants';
 
 // Make axios aware of the base path.
 axios.defaults.baseURL = window.config.BASE_URL;
@@ -22,6 +23,29 @@ const history = useRouterHistory(createHistory)({
 const store = configureStore([ routerMiddleware(history) ], { });
 const reduxHistory = syncHistoryWithStore(history, store);
 
+store.subscribe(() => {
+ switch (store.getState().lastAction.type) {
+   case constants.FETCH_SETTINGS_FULFILLED:
+     const  data = store.getState().settings.get('record');
+     const settings = data.get('settings');
+     const title = settings.get('title');
+     if(title !== '') document.title = title;
+       const css = settings.get('css');
+       if (css !== '') {
+         var head = document.getElementsByTagName('head')[0];
+         var link = document.createElement('link');
+         link.id = 'custom_css';
+         link.rel = 'stylesheet';
+         link.type = 'text/css';
+         link.href = css;
+         link.media = 'all';
+         head.appendChild(link);
+         }
+     break;
+   default:
+     break;
+   }
+});
 store.dispatch(loadCredentials());
 
 // Render application.
